Test/requirements.txt
playwright==1.40.0
playwright-stealth==1.0.6
aiohttp==3.9.1
asyncio-throttle==1.0.2
pydantic==2.5.0
apscheduler==3.10.4
sqlalchemy==2.0.23
click==8.1.7
pyyaml==6.0.1
pandas==2.1.4
fake-useragent==1.4.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
psutil==5.9.6







